UIU-879: Fix tests.
